Remove post processors and introduce transformers
